backend/requirements.txt
# --- FastAPI & Core ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1

# --- Database ---
sqlalchemy==2.0.29
asyncpg==0.30.0
alembic==1.13.1

# --- PostgreSQL Extensions ---
pgvector==0.2.4
# Switched to source version for production (requires libpq-dev in Dockerfile)
psycopg2-binary==2.9.9

# --- Authentication ---
# Note: You generally only need pyjwt OR python-jose.
# python-jose is essentially abandoned. If possible, migrate to pyjwt only.
pyjwt==2.8.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
google-auth==2.27.0

# --- Storage ---
b2sdk==1.31.0
aioboto3==13.0.0

# --- Image Processing ---
# Note: Requires libvips-dev installed in Docker
pyvips==2.2.1
imagehash==4.3.1
timm
pillow-avif-plugin
reverse_geocoder==1.5.1
piexif==1.1.3

# --- ML & Face Recognition ---
# Note: face_recognition requires cmake and dlib (slow build)
numpy==1.26.4
scikit-learn>=1.4.0
face_recognition==1.3.0

# --- AI / Object Detection ---
transformers
torch
torchvision

# --- Task Queue ---
celery==5.3.6
redis==5.0.1

# --- Utilities ---
python-dotenv==1.0.0
httpx==0.26.0
requests==2.31.0
structlog==24.1.0
prometheus-client==0.19.0

# --- Testing ---
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0